define unique_id
Increasingly more and more Home Assistant features are built on top of
entities with unique IDs.

This change defines the unique_id to be the mac to meet the requirements
of being unique within a domain.

Alternatively, we also allow users to customize it following the default
HA conventions, if they so choose.