support for configuring through the UI
This change adds a config flow so that the integration can be setup
through the UI.

For a start, we keep the old platform setup code as is for now, since I
haven't implemented an import flow.

Beyond that, this is a pretty heavy copy and paste job from other
components since the docs are pretty vague on how this stuff is meant to
work so it seems theres like 5 ways of doing everything.

I've tested reasonably well in a devcontainer with the emulated device.